test: add end-to-end test for disconnectRevokedClients (todo s3.4 phase 3)
Extract server_clients to src/server-clients.hpp so tests can include it
without pulling in the server main(). Add crl_revocation_test.cpp which
creates a real TLS server/client pair, wraps the server-side connection in
an fss_client managed by server_clients, then calls disconnectRevokedClients
with a CRL that revokes the client cert and asserts the client observes
message_type_closed. The [WARN] log line confirms the code path through
isPeerCertRevoked → disconnect → clientDisconnected is exercised end-to-end.
